chore(query-editor): move component
